+++ src/c_console.c
@@ -949,7 +949,7 @@ void C_Drawer(void)
         consoleactive = (consoledirection == 1);
 
         // cancel any gamepad vibrations
-        if (!prevconsoleactive && (gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
+        if (!prevconsoleactive && (gp_vibrate_barrels || gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
         {
             if (consoleactive)
             {