+++ src/m_menu.c
@@ -3426,7 +3426,7 @@ void M_StartControlPanel(void)
 
     S_StopSounds();
 
-    if ((gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
+    if ((gp_vibrate_barrels || gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
     {
         restoremotorspeed = idlemotorspeed;
         idlemotorspeed = 0;
@@ -3603,7 +3603,7 @@ static void M_ClearMenus(void)
 {
     menuactive = false;
 
-    if ((gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
+    if ((gp_vibrate_barrels || gp_vibrate_damage || gp_vibrate_weapons) && vibrate)
     {
         idlemotorspeed = restoremotorspeed;
         XInputVibration(idlemotorspeed);